new update of Courrier

+++ src/main/java/com/courrierpro/repositories/CourrierDepartRepository.java
@@ -4,5 +4,5 @@ import com.courrierpro.entities.CourrierDepart;
 import com.courrierpro.entitiesDTO.CourrierDepartDTO;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CourrierDepartRepository extends CrudRepository<CourrierDepart, Integer> {
+public interface CourrierDepartRepository extends CrudRepository<CourrierDepart, Long> {
 }
