+++ src/main/java/com/courrierpro/repositories/CourrierReponseRepository.java
@@ -4,5 +4,5 @@ import com.courrierpro.entities.CourrierReponse;
 import com.courrierpro.entitiesDTO.CourrierReponseDTO;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CourrierReponseRepository extends CrudRepository<CourrierReponse, Integer> {
+public interface CourrierReponseRepository extends CrudRepository<CourrierReponse, Long> {
 }
